+++ packages/gateway/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @computesdk/gateway
 
+## 0.0.4
+
+### Patch Changes
+
+- @computesdk/provider@1.0.10
+
 ## 0.0.3
 
 ### Patch Changes