+++ packages/vercel/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @computesdk/vercel
 
+## 1.6.16
+
+### Patch Changes
+
+- Updated dependencies [0c58ba9]
+  - computesdk@1.14.0
+  - @computesdk/provider@1.0.10
+
 ## 1.6.15
 
 ### Patch Changes